Read categories in from Firebase, update CategoriesAdapter
